tests: support running via pytests
adds support for running via pytest while keeping unittest functionality intact.

Due to limitations of parametrized fixture support of unittest.TestCase in
pytest, the TestCase class has to be replaced with a plain object when running
via pytest. This allows to parametrize the backend fixture for testing against
different backends.

All but 7 tests of the existing test suite require a special usb device to
be present. And this patch is not verified against running with that device
connected.